+++ src/layout.rs
@@ -1,16 +1,19 @@
 use crate::keyboard_hook::KeyAction;
 
-/// Byte 0 of [`Key::action`] contains the virtual key
+/// Byte 0 of [`Key::action`] contains the virtual key.
 const TAG_VIRTUAL_KEY: u8 = 0;
 
-/// [`Key::action`] are the bytes of a unicode code point
+/// [`Key::action`] are the bytes of a unicode code point.
 const TAG_CHARACTER: u8 = 1;
 
-/// Byte 3 of [`Key::action`] contains the target layer
+/// Changes the active layer. Byte 3 of [`Key::action`] contains the target layer.
 const TAG_MODIFIER: u8 = 2;
 
+/// Locks the current layer. Byte 3 of [`Key::action`] contains the target layer.
+const TAG_LAYER_LOCK: u8 = 3;
+
 /// Compact representation of a key action.
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 struct Key {
     scan_code: u16,
     layer: u8,
@@ -35,7 +38,7 @@ impl Key {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Default)]
 pub struct LayoutBuilder {
     keys: Vec<Key>,
     layer_names: Vec<String>,
@@ -43,10 +46,7 @@ pub struct LayoutBuilder {
 
 impl LayoutBuilder {
     pub fn new() -> Self {
-        Self {
-            keys: Vec::new(),
-            layer_names: Vec::new(),
-        }
+        Self::default()
     }
 
     fn add_or_get_layer(&mut self, layer: &str) -> u8 {
@@ -86,12 +86,28 @@ impl LayoutBuilder {
         self
     }
 
+    pub fn add_layer_lock(
+        &mut self,
+        scan_code: u16,
+        layer: &str,
+        target_layer: &str,
+        vk: Option<u8>,
+    ) -> &mut Self {
+        let layer = self.add_or_get_layer(layer);
+        let target_layer = self.add_or_get_layer(target_layer);
+        let key = Key {
+            scan_code,
+            layer,
+            tag: TAG_LAYER_LOCK,
+            action: [vk.unwrap_or(0), 0, 0, target_layer],
+        };
+        self.keys.push(key);
+        self
+    }
+
     pub fn build(mut self) -> Layout {
         self.keys.sort_by_key(|k| k.scan_code);
-        Layout {
-            keys: self.keys,
-            layer_names: self.layer_names,
-        }
+        Layout { keys: self.keys }
     }
 }
 
@@ -105,7 +121,6 @@ pub struct Modifier {
 #[derive(Debug)]
 pub struct Layout {
     keys: Vec<Key>,
-    layer_names: Vec<String>,
 }
 
 impl Layout {
@@ -130,6 +145,17 @@ impl Layout {
                 layer_to: k.action[3],
             })
     }
+
+    pub fn layer_locks(&self) -> impl Iterator<Item = Modifier> + '_ {
+        self.keys
+            .iter()
+            .filter(|k| k.tag == TAG_LAYER_LOCK)
+            .map(|k| Modifier {
+                scan_code: k.scan_code,
+                layer_from: k.layer,
+                layer_to: k.action[3],
+            })
+    }
 }
 
 pub struct KeyResults<'a> {
@@ -139,7 +165,7 @@ pub struct KeyResults<'a> {
 }
 
 impl<'a> KeyResults<'a> {
-    pub fn action_on_layer(&self, layer: u8) -> Option<KeyAction> {
+    fn key_on_layer(&self, layer: u8) -> Option<Key> {
         let iter_back = self.keys[..self.idx]
             .iter()
             .rev()
@@ -147,11 +173,16 @@ impl<'a> KeyResults<'a> {
         let iter_forward = self.keys[self.idx..]
             .iter()
             .take_while(|k| k.scan_code == self.scan_code);
-        let mut iter = iter_back.chain(iter_forward);
+        iter_back
+            .chain(iter_forward)
+            .find(|k| k.layer == layer)
+            .copied()
+    }
 
-        let key = iter.find(|k| k.layer == layer)?;
+    pub fn action(&self, layer: u8) -> Option<KeyAction> {
+        let key = self.key_on_layer(layer)?;
         let action = match key.tag {
-            TAG_VIRTUAL_KEY | TAG_MODIFIER => match key.action[0] {
+            TAG_VIRTUAL_KEY | TAG_MODIFIER | TAG_LAYER_LOCK => match key.action[0] {
                 0 => KeyAction::Ignore,
                 vk => KeyAction::VirtualKey(vk),
             },
@@ -161,5 +192,13 @@ impl<'a> KeyResults<'a> {
             _ => unreachable!(),
         };
         Some(action)
+    }
+
+    pub fn layer_lock(&self, layer: u8) -> Option<u8> {
+        let key = self.key_on_layer(layer)?;
+        match key.tag {
+            TAG_LAYER_LOCK => Some(key.action[3]),
+            _ => None,
+        }
     }
 }